refactor: consolidate database population scripts to /scripts/populate/
- Moved populate_play_by_play.py and populate_player_season_stats.py
  to /scripts/populate/ with refactoring to use shared infrastructure
- Fixed import paths in existing populate package (populate_db -> populate)
- Removed duplicate scripts from scripts root
- Updated __init__.py to export all population modules
- Refactored scripts to use shared NBAClient, config, and logging
- Added reference/nba_api/ to .gitignore for cloned nba_api docs

The populate package now provides a unified interface for all database
population operations with consistent rate limiting, retry logic,
progress tracking, and error handling.
